requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-rest-knox==4.2.0
djangorestframework==3.15.1
Markdown==3.6
packaging==24.1
Pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
shortuuid==1.0.13
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
waitress==3.0.0
whitenoise==6.9.0
